Added files for database interaction
Since I did not want the croud my computer with a postgres server like a
pleb, there is a docker compose file included which will stand up a
virtual instance of a postgres server.

Once the server is up, the db_interface file is used to send commands to
the postgres server. So far it is just a bare bone structure with table
create/destroy commands.